Refactoring ExamController Finished
