+++ OnlineExam/Controllers/ExamController.cs
@@ -2,6 +2,7 @@
 using Microsoft.AspNetCore.Mvc;
 using OnlineExam.Data;
 using OnlineExam.Models;
+using OnlineExam.Repository.DashboardRepo;
 using OnlineExam.Repository.ExamRepos;
 using OnlineExam.Repository.QuestionsRepo;
 using OnlineExam.ViewModels;
@@ -12,24 +13,21 @@ namespace OnlineExam.Controllers
     [Authorize]
     public class ExamController : Controller
     {
-        private readonly ApplicationDbContext _context;
         private readonly IExamRepo _examContext;
         private readonly IQuestionRepo _questionContext;
-        public ExamController(ApplicationDbContext context, IExamRepo examContext, IQuestionRepo questionContext)
+        private readonly IDashboardRepo _dashboardContext;
+        public ExamController(IExamRepo examContext, IQuestionRepo questionContext, IDashboardRepo dashboardContext)
         {
-            _context = context;
             this._examContext = examContext;
             this._questionContext = questionContext;
+            this._dashboardContext = dashboardContext;
         }
-
 
         public IActionResult Index()
         {
-
             var userId = this.User.FindFirstValue(ClaimTypes.NameIdentifier);
 
             var Examlist = _examContext.GetAllExams(userId);
-
 
             return View(Examlist);
         }
@@ -50,14 +48,10 @@ namespace OnlineExam.Controllers
 
         public IActionResult EditExam(int Id)
         {
-
             var item = _examContext.FindExam(Id);
 
             if (item == null)
-            {
                 return NotFound();
-            }
-
 
             return View(item);
         }
@@ -70,14 +64,12 @@ namespace OnlineExam.Controllers
             return RedirectToAction("Index");
         }
 
+        [HttpGet]
         public IActionResult Delete(int Id)
         {
-            if (Id == 0)
-            {
+            var remove = _examContext.RemoveExam(Id);
+            if(remove == false)
                 return NotFound();
-            }
-
-            _examContext.RemoveExam(Id);
 
             return RedirectToAction("Index");
         }
@@ -126,7 +118,6 @@ namespace OnlineExam.Controllers
         [HttpGet]
         public IActionResult DeleteQuestion(int Id)
         {
-
             var question = _questionContext.FindQuestion(Id);
 
             int id = question.ExamId;
@@ -136,8 +127,7 @@ namespace OnlineExam.Controllers
             return RedirectToAction("ViewQuestions", new { Id = id });
         }
 
-
-
+        [HttpGet]
         public IActionResult EditQuestion(int Id)
         {
             var question = _questionContext.FindQuestion(Id);
@@ -158,169 +148,32 @@ namespace OnlineExam.Controllers
             return RedirectToAction("ViewQuestions", new { Id = item.ExamId });
         }
 
-
+        [HttpGet]
         public IActionResult Dashboard()
         {
-            List<DashboardViewModel> model = new List<DashboardViewModel>();
+            var dashboards = _dashboardContext.Dashboards();
 
-            foreach (var i in _context.Answers)
-            {
-                DashboardViewModel item = new DashboardViewModel
-                {
-                    StudentId = i.StudentNationalId,
-                    StudentName = i.StudentName,
-                    Score = (int)i.Score,
-                    ExamName = _context.Exams.First(j => j.ExamId == i.ExamId).Title
-
-                };
-
-                model.Add(item);
-
-            }
-
-
-            return View(model);
+            return View(dashboards);
         }
 
-        public IActionResult ViewResults(int? id)
+        [HttpGet]
+        public IActionResult ViewResults(int id)
         {
-            if (id == null || id == 0)
-            {
+            var model = _dashboardContext.GetResult(id);
+            if(model is null)
                 return NotFound();
-            }
-
-            ResultsModel model = new ResultsModel();
-            model.ExamName = _context.Exams.Where(i => i.ExamId == id).ToArray()[0].Title;
-
-            model.answers = _context.Answers.Where(i => i.ExamId == id).ToList();
-
-            int total = model.answers.Count;
-            int numOFQuestions = _context.Questions.Where(i => i.ExamId == id).Count();
-            int critical = numOFQuestions / 2;
-            int numOfSucceeded = 0, sumOfScores = 0;
-            model.isStudentPassed = new List<bool>();
-            model.StudentsStatus = new List<String>();
-
-            foreach (var i in model.answers)
-            {
-                if (i.Score >= critical)
-                {
-                    numOfSucceeded++;
-                    model.isStudentPassed.Add(true);
-
-                }
-                else
-                {
-                    model.isStudentPassed.Add(false);
-                }
-                model.StudentsStatus.Add(GetLetterGrade((int)i.Score, total));
-                sumOfScores += (int)i.Score;
-
-            }
-
-            model.numberOfPassedStudent = numOfSucceeded;
-            model.numberOfFailedStudent = total - numOfSucceeded;
-            total = Math.Max(total, 1);
-            model.averageGrade = sumOfScores / total;
 
             return View(model);
         }
-        public string GetLetterGrade(int totalScore, int totalQuestions)
-        {
-            double scorePercent = ((double)totalScore / totalQuestions) * 100;
-            string letterGrade;
 
-            if (scorePercent >= 90)
-            {
-                letterGrade = "A+";
-            }
-            else if (scorePercent >= 85)
-            {
-                letterGrade = "A";
-            }
-            else if (scorePercent >= 80)
-            {
-                letterGrade = "A-";
-            }
-            else if (scorePercent >= 75)
-            {
-                letterGrade = "B+";
-            }
-            else if (scorePercent >= 70)
-            {
-                letterGrade = "B";
-            }
-            else if (scorePercent >= 65)
-            {
-                letterGrade = "B-";
-            }
-            else if (scorePercent >= 60)
-            {
-                letterGrade = "C+";
-            }
-            else if (scorePercent >= 55)
-            {
-                letterGrade = "C";
-            }
-            else if (scorePercent >= 50)
-            {
-                letterGrade = "C-";
-            }
-            else if (scorePercent >= 45)
-            {
-                letterGrade = "D+";
-            }
-            else if (scorePercent >= 40)
-            {
-                letterGrade = "D";
-            }
-            else
-            {
-                letterGrade = "F";
-            }
-
-            return letterGrade;
-        }
-        public IActionResult ExamResultDetails(int? id)
+        [HttpGet]
+        public IActionResult ExamResultDetails(int id)
         {
-            if (id == null || id == 0)
-            {
+            var model = _dashboardContext.ExamDetails(id);
+            if(model is null)
                 return NotFound();
-            }
-
-            ExamAnswerResult model = new ExamAnswerResult();
-
-            Answer ans = _context.Answers.Where(i => i.AnswerId == id).ToList()[0];
-            model.StudentName = ans.StudentName;
-            int ExamId = _context.Answers.Where(i => i.AnswerId == id).ToList()[0].ExamId;
-            model.ExamName = _context.Exams.Where(i => i.ExamId == ExamId).ToList()[0].Title;
-
-            model.lstOfQuestionAnswers = _context.AnswerQuestions.Where(i => i.AnswerId == id).ToList();
-
-            model.numOfQuestions = _context.Questions.Where(i => i.ExamId == ExamId).Count();
-
-            int score = 0;
-
-            foreach (var i in model.lstOfQuestionAnswers)
-            {
-                if (i.TrueAnswer == i.SelectedAnswer)
-                    score++;
-
-
-            }
-
-            model.score = score;
-
-            if (score * 2 >= model.numOfQuestions)
-                model.passed = true;
-
-
-
 
             return View(model);
         }
-
-
-
     }
 }
