+++ OnlineExam/ViewModels/ExamAnswerResult.cs
@@ -4,23 +4,12 @@ namespace OnlineExam.ViewModels
 {
     public class ExamAnswerResult
     {
-        public List<AnswerQuestion> lstOfQuestionAnswers {  get; set; }
-
+        public int StudentId { set; get; }
         public string ExamName { set; get; }
         public string StudentName { set;get; }
-        public int StudentId { set; get; }
-
         public int numOfQuestions { set; get; }
         public int score { set; get; }
         public bool passed { set; get; }=false;
-        
-
+        public List<AnswerQuestion> lstOfQuestionAnswers {  get; set; }
     }
 }
-//start time exam
-//name of student
-//id of student
-//score
-//boolean passed or not
-//grade
-//number of questions
