+++ OnlineExam/Views/Exam/ExamResultDetails.cshtml
@@ -142,106 +142,7 @@
         }
     </div>
 </div>
-<style>
-    .question-container {
-        padding: 16px 16px 16px 32px;
-        background: #D9D9D9;
-        border-radius: 18px;
-        margin-bottom: 24px;
-        width: 65%;
-    }
 
-    .question-txt-container p {
-        font-size: 16px;
-        font-weight: 600;
-        margin: 0;
-        padding: 12px;
-    }
-
-    .question-container-wal-marah-de-mgd {
-        gap: 30px;
-        display: grid;
-    }
-
-    .question-txt-container {
-        background-color:white;
-        width: 100%;
-        display: flex;
-        justify-content: space-between;
-        border-radius: 18px;
-        align-items: center;
-        border: 1px solid;
-    }
-
-
-    .answer-container {
-        display: flex;
-        align-items: center;
-        gap: 20px;
-    }
-
-    .icon {
-        color: #000EE6;
-        margin-right: 16px;
-    }
-
-    .option-container {
-        width: 100%;
-        margin-top: 12px;
-        gap: 37px;
-        display: inline-flex;
-        justify-content: flex-end;
-        align-items: center;
-    }
-
-    .header-container {
-        display: flex;
-        justify-content: space-between;
-        padding-top: 18px;
-        padding-bottom: 18px;
-        padding-left: 72px;
-        padding-right: 72px;
-        box-shadow: -3px 1px 5px 0px rgba(0, 0, 255, .2);
-    }
-
-        .header-container p {
-            font-size: 18px;
-            font-weight: 700;
-            margin: 0px;
-        }
-
-    .studentInfo-container {
-        display: flex;
-        width: 50%;
-        flex-direction: row;
-        align-items: center;
-        justify-content: space-between;
-        padding: 24px;
-        margin-left: 320px;
-        border-radius: 12px 12px 12px 12px;
-        box-shadow: 2px 3px 5px 0px rgba(0, 0, 255, .2);
-        margin-top: 42px;
-        margin-bottom: 42px;
-        padding-left: 82px;
-        padding-right: 82px;
-    }
-
-    .Questions-container {
-        display: flex;
-        flex-direction: column;
-        flex-wrap: nowrap;
-        align-items: center;
-    }
-
-    .questionHeader-container {
-        display: flex;
-        justify-content: space-between;
-        padding-left: 32px;
-        padding-right: 32px;
-        padding-bottom: 12px;
-    }
-    .studentInfo-container p{
-        font-size:16px;
-        font-weight:bold;
-    }
-</style>
+@section PageStyle {
+    <link rel="stylesheet" href="~/css/ExamResultDetails.css">  
+}
