+++ OnlineExam/wwwroot/css/ExamViewResults.css
@@ -1,0 +1,68 @@
+.examName-container {
+    padding: 48px 12px 12px 88px;
+}
+
+.examName-container p {
+    font-size: 20px;
+    font-weight: 700;
+}
+
+.examInfo-container {
+    display: flex;
+    flex-direction: row;
+    align-items: center;
+    justify-content: center;
+    gap: 78px;
+    width: 70%;
+    padding: 24px;
+    margin-left: 206px;
+    border-radius: 12px 12px 12px 12px;
+    box-shadow: 2px 3px 5px 0px rgba(0, 0, 255, .2);
+}
+
+.totalStudent-container {
+    display: flex;
+    align-items: center;
+    gap: 14px;
+    border-right: 1px solid;
+    padding-right: 8px;
+}
+
+.examInfo-container p {
+    margin: 0px;
+    font-size: 14px;
+    font-weight: 600;
+}
+
+.Info-container {
+    display: flex;
+    gap: 10px;
+    flex-direction: column;
+    align-items: center;
+}
+
+.label-container {
+    display: flex;
+    background: #D9D9D9;
+    flex-direction: row;
+    padding: 18px 64px 17px 58px;
+    justify-content: space-between;
+    align-items: center;
+}
+
+.label-container p {
+    font-weight: bolder;
+    margin: 0px;
+}
+
+.result-container {
+    display: flex;
+    padding-left: 64px;
+    border-bottom: 1px solid;
+    padding-top: 12px;
+    padding-bottom: 12px;
+}
+
+.examResults-container {
+    margin-top: 42px;
+}
